feat: 秒杀服务搭建、基础配置实现
获取当前时间段秒杀时间段秒杀商品
获取商品秒杀信息

+++ gulimall-coupon/src/main/java/com/laughingather/gulimall/coupon/entity/SeckillSkuRelationEntity.java
@@ -37,20 +37,20 @@ public class SeckillSkuRelationEntity implements Serializable {
 	 */
 	private Long skuId;
 	/**
-	 * 秒杀价格
-	 */
-	private BigDecimal seckillPrice;
-	/**
-	 * 秒杀总量
-	 */
-	private BigDecimal seckillCount;
-	/**
-	 * 每人限购数量
-	 */
-	private BigDecimal seckillLimit;
-	/**
-	 * 排序
-	 */
-	private Integer seckillSort;
+     * 秒杀价格
+     */
+    private BigDecimal seckillPrice;
+    /**
+     * 秒杀总量
+     */
+    private Integer seckillCount;
+    /**
+     * 每人限购数量
+     */
+    private Integer seckillLimit;
+    /**
+     * 排序
+     */
+    private Integer seckillSort;
 
 }
